requirements.txt
# Core dependencies
httpx>=0.24.0
asyncio
python-dotenv>=1.0.0

# AgentMail SDK
agentmail>=0.1.0

# Browser automation (optional - only for account creation)
playwright>=1.40.0
playwright-stealth>=2.0.0

# Image processing (for mosaic)
Pillow>=10.0.0

# Optional utilities
aiofiles>=23.0.0
